requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.110.0
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.1
idna==3.6
multidict==6.0.5
packaging==24.0
playwright==1.42.0
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
tokenizers==0.15.2
tqdm==4.66.2
twilio==9.0.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
